Retry on error while fetching inbox notifications
